pyproject.toml
[project]
name = "khc-cli"
version = "0.5.1"
description = "CLI tool for analyzing and curating awesome lists focused on sustainable and open source technologies"
readme = "README.md"
authors = [
    { name = "Abderrahim Guennoune", email = "aguennoune@outlook.com" }
]
requires-python = ">=3.10,<4.0"
license = { file = "LICENSE" }
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License", 
    "Operating System :: OS Independent",
    "Environment :: Console",
    "Topic :: Utilities",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Information Technology",
]
keywords = ["cli", "awesome-list", "github", "sustainability", "open-source", "technology", "curation", "analysis"]
dependencies = [
    "typer[all]>=0.9.0,!=0.16.0",
    "rich>=13.6.0",
    "pygithub>=2.1.1",
    "beautifulsoup4>=4.12.2",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "termcolor>=2.3.0",
    "typing-extensions>=4.8.0",
    "markdown>=3.8",
    "twine>=6.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "ruff>=0.0.270",
    "mypy>=1.3.0",
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=1.2.0",
    "sphinx-click>=4.4.0",
]
docs = [
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=1.2.0",
    "sphinx-click>=4.4.0",
]

[project.urls]
"Homepage" = "https://github.com/Krypto-Hashers-Community/khc-cli"
"Bug Tracker" = "https://github.com/Krypto-Hashers-Community/khc-cli/issues"
"Documentation" = "https://khc-cli.readthedocs.io/"

[project.scripts]
khc-cli = "khc_cli.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.targets.wheel]
include = [
    "src/khc_cli/resources/*.md",
]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I"]
ignore = []

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"